feat: add authored Aurelia Magna city and siege mission
Introduce Aurelia Magna as a single authored city definition shared by the arena presentation and the siege mission. Ship the generated map and mission assets, register them with the resource and translation catalogs, and preserve the city layout, terrain, roads, rivers, bridges, structures, units, patrol groups, and world props in data that can be loaded at runtime.

Extend map definitions and loading to retain authored spawn and structure groups and to support tapered, raise-only flat terrain features. Add the city export tool and its authored plan so the large map remains reproducible from explicit city geometry, terrain, route, garrison, and landmark data.

Build the arena city scenario from the shipped map instead of maintaining a second hand-authored layout. Keep the arena-specific presentation and patrol behavior layered on top of the shared map, and add parity coverage for city counts, landmark positions, routes, ownership, and mission separation.

Update the related mission, reachability, commander-control, portrait, rendering, and hill-shape integration points so the new city assets participate in existing loading and presentation flows.
