+++ assets.qrc
@@ -164,6 +164,7 @@
         <file>assets/maps/map_mountain.json</file>
         <file>assets/maps/map_copper_canyons.json</file>
         <file>assets/maps/map_amber_delta.json</file>
+        <file>assets/maps/map_aurelia_magna.json</file>
         <file>assets/maps/map_tutorial.json</file>
         <!-- Only git-tracked campaign map files belong here. The generated
              tools/map_pipeline outputs (base color, water, coastlines_uv,
@@ -185,6 +186,7 @@
         <file>assets/missions/iron_sepulcher_watch.json</file>
         <file>assets/missions/the_timber_levy.json</file>
         <file>assets/missions/hold_the_sallow_ford.json</file>
+        <file>assets/missions/siege_of_aurelia_magna.json</file>
         <file>assets/missions/tutorial.json</file>
 
         <!-- Audio assets. Music and ambience are deliberately absent:
