+++ tools/CMakeLists.txt
@@ -3,6 +3,7 @@ add_subdirectory(balance_sim)
 add_subdirectory(map_editor)
 add_subdirectory(content_validator)
 add_subdirectory(terrain_probe)
+add_subdirectory(city_export)
 add_subdirectory(movement_trace_report)
 add_subdirectory(bpat_baker)
 add_subdirectory(audio_master)
