+++ tools/city_export/aurelia_magna_plan.h
@@ -1,0 +1,24 @@
+#pragma once
+
+#include <QString>
+#include <QVector3D>
+
+#include <vector>
+
+#include "tools/arena/arena_scenario.h"
+
+namespace Arena::Scenarios {
+
+struct CityPatrolRoute {
+  QString name;
+  std::vector<QVector3D> route;
+};
+
+struct CityPlan {
+  ArenaScenarioDefinition definition;
+  std::vector<CityPatrolRoute> patrols;
+};
+
+[[nodiscard]] auto authored_city_plan() -> CityPlan;
+
+} // namespace Arena::Scenarios
