+++ translations/asset_strings_generated.cpp
@@ -13,8 +13,8 @@ namespace {
 
 // clang-format off
 [[maybe_unused]] const char* const k_asset_strings[] = {
-    // 782 strings across 9 contexts.
-    // ---- Missions (304) ----
+    // 813 strings across 9 contexts.
+    // ---- Missions (333) ----
     QT_TRANSLATE_NOOP("Missions", "202 BC: Scipio arranged lanes through his infantry to blunt Hannibal's elephants, while Masinissa and Laelius commanded superior allied cavalry. The returning cavalry helped decide the battle. The Sepulcher intervention is the campaign's deliberate dark-fantasy break from history."),
     QT_TRANSLATE_NOOP("Missions", "215-212 BC: Hannibal campaigned across Campania while Capua served as his principal Italian ally and winter base. This dark-fantasy composite condenses several Roman attempts to contain and isolate the Carthaginian army."),
     QT_TRANSLATE_NOOP("Missions", "216 BC: beside the Aufidus, Hannibal's deliberately yielding center drew a much larger Roman infantry mass forward while African infantry attacked its flanks and Carthaginian cavalry closed the rear. Two Roman camps supported the army."),
@@ -33,6 +33,7 @@ namespace {
     QT_TRANSLATE_NOOP("Missions", "A guided first field, paced by the player rather than the clock: the Roman raid waits until the recruit's army stands, and the enemy camp is small enough to fall to a first attack."),
     QT_TRANSLATE_NOOP("Missions", "A nation stands on its commander. Kill him and the camps go quiet and unclaimed - which, against a twenty-minute clock, is the only version of this battle that finishes in time."),
     QT_TRANSLATE_NOOP("Missions", "A new standard in the valley, and nobody under it who has held a spear more than twice. My outpost has orders to go and look at you. Try to still be standing when they arrive; it makes the report worth writing."),
+    QT_TRANSLATE_NOOP("Missions", "A raid on the pasture, then a landing at the quay. Nine gates and one garrison: the avenue and the decumanus are how you get from one to the other."),
     QT_TRANSLATE_NOOP("Missions", "ASSAULT I: infantry from north and south; cavalry raiders and engineers on the eastern consular road."),
     QT_TRANSLATE_NOOP("Missions", "ASSAULT II: the Roman horns answer one another. Rotate the reserve before all three roads close."),
     QT_TRANSLATE_NOOP("Missions", "ASSAULT III: siege engines and veteran cohorts. Hold every gate until the twentieth minute."),
@@ -47,14 +48,17 @@ namespace {
     QT_TRANSLATE_NOOP("Missions", "Battle of Zama"),
     QT_TRANSLATE_NOOP("Missions", "Between assaults, seize the eastern Roman siege camp."),
     QT_TRANSLATE_NOOP("Missions", "Break all five columns. The ford is held when the last of them is dead in front of it, and not one column sooner."),
+    QT_TRANSLATE_NOOP("Missions", "Break all six assaults. They do not come at the same gate twice."),
     QT_TRANSLATE_NOOP("Missions", "Break all three Roman assault phases. The quarter holds when the last consular column is dead in the streets."),
     QT_TRANSLATE_NOOP("Missions", "Break both Iron Sepulcher attacks. At Zama, the dead no longer remain outside the battle."),
     QT_TRANSLATE_NOOP("Missions", "Break the chiefs contesting the descent."),
     QT_TRANSLATE_NOOP("Missions", "Break the consular line at its centre."),
+    QT_TRANSLATE_NOOP("Missions", "Break the siege"),
     QT_TRANSLATE_NOOP("Missions", "Break the siege camp once the assaults are spent."),
     QT_TRANSLATE_NOOP("Missions", "Break the three assault phases"),
     QT_TRANSLATE_NOOP("Missions", "Break the three crossings"),
     QT_TRANSLATE_NOOP("Missions", "Break three Roman columns before the survivors find the ford again."),
+    QT_TRANSLATE_NOOP("Missions", "Bring in the first harvest"),
     QT_TRANSLATE_NOOP("Missions", "Bring the column to the first pass"),
     QT_TRANSLATE_NOOP("Missions", "Bring the column to the pontoon bridge on the near bank."),
     QT_TRANSLATE_NOOP("Missions", "Burn every cursed guardian from the sepulcher ruins. Kill what cannot die."),
@@ -83,6 +87,7 @@ namespace {
     QT_TRANSLATE_NOOP("Missions", "Enemy camps fall neutral once their commander is dead."),
     QT_TRANSLATE_NOOP("Missions", "Every soldier lost is counted by the enemy tonight."),
     QT_TRANSLATE_NOOP("Missions", "Every system a first match asks for, in the order a new commander meets them: selection and orders, the resource loop from tree to yard, construction cost and placement, recruitment and the population that limits it, stances, the commander's aura and rally, camera, speed and objectives - then one small assault to put it all together."),
+    QT_TRANSLATE_NOOP("Missions", "Everything I have is on the roads, consul. East, west, and the bridge. Choose two."),
     QT_TRANSLATE_NOOP("Missions", "Execute the encirclement, then seize all three Roman camps before their reserves can restore the line."),
     QT_TRANSLATE_NOOP("Missions", "Field Training"),
     QT_TRANSLATE_NOOP("Missions", "Fifteen thousand in the shallows, and the augurs will say the sacred chickens warned us. Kill a consul and Rome elects another by market day. That is the trick of us: we are not a man."),
@@ -98,6 +103,7 @@ namespace {
     QT_TRANSLATE_NOOP("Missions", "Fix the Roman center in place. Keep both cavalry wings active."),
     QT_TRANSLATE_NOOP("Missions", "Flaminius must not leave the shore."),
     QT_TRANSLATE_NOOP("Missions", "Follow the valley floor east; the mountain peoples hold the switchbacks above."),
+    QT_TRANSLATE_NOOP("Missions", "Grain from the river fields, timber from the groves, stone from the quarries beyond the north wall. Every assault you break early is a window to work."),
     QT_TRANSLATE_NOOP("Missions", "Grave-lights move among the western orchards. The Iron Sepulcher is watching, but the legions remain the greater danger."),
     QT_TRANSLATE_NOOP("Missions", "Hannibal dies in the snow. Then the elephants were for nothing, and so was the road. History will find that very tidy."),
     QT_TRANSLATE_NOOP("Missions", "Hannibal falls at the Rhone. Then Rome never sees the elephants, and I never see Cannae. I shall spend the rest of my life not knowing what I was spared."),
@@ -115,6 +121,7 @@ namespace {
     QT_TRANSLATE_NOOP("Missions", "Hold the Sallow Ford"),
     QT_TRANSLATE_NOOP("Missions", "Hold the field, then. My horse is dead, my son is dragging me off it by the arm, and I find I can still count. You have won a ford. Italy is nine hundred miles long."),
     QT_TRANSLATE_NOOP("Missions", "Hold the gate yard of Scipio's forward town."),
+    QT_TRANSLATE_NOOP("Missions", "Hold the outer gates"),
     QT_TRANSLATE_NOOP("Missions", "Hold the river line and break all three Roman crossings. Let the cold and the water do the work Sempronius will not."),
     QT_TRANSLATE_NOOP("Missions", "Hold the ruins. Force the Romans to fight on two grounds at once."),
     QT_TRANSLATE_NOOP("Missions", "Hold the southern bank through all three crossings. Sempronius spends his army on the ice; you have only to still be standing."),
@@ -127,6 +134,7 @@ namespace {
     QT_TRANSLATE_NOOP("Missions", "It is the closest of the four camps to your line."),
     QT_TRANSLATE_NOOP("Missions", "Keep Capua, then. I shall come back after the harvest, and after the one following it. I have buried three consuls who were in a hurry. Patience is cheaper, and it does not require a triumph."),
     QT_TRANSLATE_NOOP("Missions", "Keep both towers standing. Lose one and the lane it watches is yours to walk yourself."),
+    QT_TRANSLATE_NOOP("Missions", "Keep every hall. The two granges and the two quarry camps stand outside the wall, and they are the only stockpiles the work crews can reach."),
     QT_TRANSLATE_NOOP("Missions", "Kill Sempronius and his officers"),
     QT_TRANSLATE_NOOP("Missions", "Kill every enemy commander. A nation dies with the man who leads it: its camps fall neutral, its works come down and its troops leave the field."),
     QT_TRANSLATE_NOOP("Missions", "Kill the Roman commander and take the outpost. A nation dies with the man who leads it."),
@@ -138,6 +146,8 @@ namespace {
     QT_TRANSLATE_NOOP("Missions", "Legionaries at your works. Not Varro's legions; mine. Men who were at Cannae and remember exactly how it felt."),
     QT_TRANSLATE_NOOP("Missions", "Levy filled. Stack it by the road and let the carts argue about it. We cut a wood down to feed a siege that has not started yet — remember that the next time somebody calls this the quiet posting."),
     QT_TRANSLATE_NOOP("Missions", "Light is going. Five minutes of road left in the day, and whatever is on the carts when the column reaches us is what we levied. Bring them home."),
+    QT_TRANSLATE_NOOP("Missions", "Look at it. Three walls, seventeen gates, and not one stalk of grain growing inside any of them. The city was built to be admired, not besieged. Get the crews over the river while the road is still ours, and keep the garrison off the ramparts until you know which gate they want."),
+    QT_TRANSLATE_NOOP("Missions", "Lose every hall in the city and there is nothing left to recruit from, or carry a harvest to."),
     QT_TRANSLATE_NOOP("Missions", "Lose the redoubt and there is nothing behind it for forty miles."),
     QT_TRANSLATE_NOOP("Missions", "Lose the river camp and the cold takes what the Romans couldn't."),
     QT_TRANSLATE_NOOP("Missions", "Lose the supply camp and the army starves before it reaches the snow line."),
@@ -154,17 +164,21 @@ namespace {
     QT_TRANSLATE_NOOP("Missions", "No living formation remains to hold the gates."),
     QT_TRANSLATE_NOOP("Missions", "Nothing new is taught here. The ridge, the bait and the beheading are all things you already know; Cannae only asks whether you can run three of them at once against four times your number."),
     QT_TRANSLATE_NOOP("Missions", "Nothing on the shore road survives once both ends are held."),
+    QT_TRANSLATE_NOOP("Missions", "One consul and an empty city is not a garrison."),
     QT_TRANSLATE_NOOP("Missions", "One minute. Spend the reserve; save the camp."),
     QT_TRANSLATE_NOOP("Missions", "One officer, no line. The next column will not even slow down."),
     QT_TRANSLATE_NOOP("Missions", "One redoubt, two mouths, five columns. Break every assault that comes down the cart gate and the goat path, and still be standing when the fifth is spent."),
     QT_TRANSLATE_NOOP("Missions", "Overcut the levy. A surplus is the only thing a quartermaster has ever thanked anyone for."),
     QT_TRANSLATE_NOOP("Missions", "Patrols at your works. You are trying to cross a river with your back to my roads. It is bold. It is also how one drowns."),
     QT_TRANSLATE_NOOP("Missions", "Preserve the market and four homes: Campania is being defended for its people and supplies, not merely its walls."),
+    QT_TRANSLATE_NOOP("Missions", "Provision the capital"),
+    QT_TRANSLATE_NOOP("Missions", "Provision the capital. The grain is across the river and the stone is past the quarry gate; neither will walk in by itself."),
     QT_TRANSLATE_NOOP("Missions", "Provision the column for the descent: harvest timber, stone and iron from the pass while the mountain tribes contest every switchback."),
     QT_TRANSLATE_NOOP("Missions", "Raze every Roman post. Leave Scipio nothing to rally behind."),
     QT_TRANSLATE_NOOP("Missions", "Reach the Rhone crossing"),
     QT_TRANSLATE_NOOP("Missions", "Rear guard deployment detected across the southern wadi. This is the final phase."),
     QT_TRANSLATE_NOOP("Missions", "Ride down the commanders once the crossings break."),
+    QT_TRANSLATE_NOOP("Missions", "Riders on the pasture gate. They are not here for the wall; they are here for the flocks. Whatever you send has to get there before the herds do."),
     QT_TRANSLATE_NOOP("Missions", "Roman cavalry is attempting a breakout from the south. Intercept them."),
     QT_TRANSLATE_NOOP("Missions", "Roman cavalry screens the field. Punch through before reserves arrive."),
     QT_TRANSLATE_NOOP("Missions", "Roman cavalry waves intensify. Collapse one flank before the rear guard arrives."),
@@ -190,6 +204,8 @@ namespace {
     QT_TRANSLATE_NOOP("Missions", "Sempronius is crossing. Now, in the cold, without breakfast, because a consul's year ends in March and he wants his battle before it does."),
     QT_TRANSLATE_NOOP("Missions", "Sempronius spends his army on the ice; you only have to still be standing."),
     QT_TRANSLATE_NOOP("Missions", "Set builders on the slopes; the pass is paid for in materials, not blood."),
+    QT_TRANSLATE_NOOP("Missions", "Six assaults and the granaries full. The city will say the walls held. The walls did nothing -- the road to the river held, and the crews who walked it did not stop walking it. Write that down and let them carve the other version."),
+    QT_TRANSLATE_NOOP("Missions", "Six assaults, and the last one comes at three gates with elephants behind it."),
     QT_TRANSLATE_NOOP("Missions", "Sixty men and a shrine the local people will not walk past after dark. My orders say investigate. I would sooner burn it and write that we investigated, but the Senate reads."),
     QT_TRANSLATE_NOOP("Missions", "So he dies by his own lake in his own fog. Rome will call it justice. I call it weather, and I have always respected weather."),
     QT_TRANSLATE_NOOP("Missions", "So the centre gave and the wings did not. That was well done, and I say so, and I was inside it. Rome has lost an army and a ring for every knight in it. Rome has not yet lost anything that matters."),
@@ -200,6 +216,7 @@ namespace {
     QT_TRANSLATE_NOOP("Missions", "Spring the ambush and capture both Roman field camps within twenty minutes, before the mist lifts and a cohort escapes the basin."),
     QT_TRANSLATE_NOOP("Missions", "Stand in the gate yard of the northern advance camp until it falls to Carthage."),
     QT_TRANSLATE_NOOP("Missions", "Stand your ground. The last soldier alone is not a force — he is a sacrifice."),
+    QT_TRANSLATE_NOOP("Missions", "Stone is the scarce one: the only boulders are at the two quarries outside the walls."),
     QT_TRANSLATE_NOOP("Missions", "Strike now. Close the western route and split their line."),
     QT_TRANSLATE_NOOP("Missions", "Supply is a target - and it cuts both ways. Your market and your homes are what the Romans are actually marching at, and their siege camp keeps its own stockpile within reach of a raid."),
     QT_TRANSLATE_NOOP("Missions", "Take Scipio's town"),
@@ -218,6 +235,7 @@ namespace {
     QT_TRANSLATE_NOOP("Missions", "Take the second camp on the southern supply road."),
     QT_TRANSLATE_NOOP("Missions", "Take the shore camp"),
     QT_TRANSLATE_NOOP("Missions", "Taking the rear first denies the cavalry a place to rally."),
+    QT_TRANSLATE_NOOP("Missions", "That gate holds. Do not stand there admiring it -- the crews go out now, and they come back before the next horn."),
     QT_TRANSLATE_NOOP("Missions", "The Aufidus camp. So the ring closes. I told the Senate a ring was what he wanted. They told me I was old."),
     QT_TRANSLATE_NOOP("Missions", "The Barcid falls at the Ticinus. Then this was the whole war, and it fit inside an afternoon. I expected more of it; I am not sure I wanted more."),
     QT_TRANSLATE_NOOP("Missions", "The Campanian Vigil"),
@@ -226,6 +244,7 @@ namespace {
     QT_TRANSLATE_NOOP("Missions", "The Roman column walks into a geographical trap and is destroyed by converging attacks while the lakeside Sepulcher altar remains an optional horror."),
     QT_TRANSLATE_NOOP("Missions", "The Roman villages are heavily defended. Coordinate your attacks!"),
     QT_TRANSLATE_NOOP("Missions", "The Romans are baited into the winter river and punished by Mago's concealed force, then the defense becomes an assault on a complete winter camp."),
+    QT_TRANSLATE_NOOP("Missions", "The Siege of Aurelia Magna"),
     QT_TRANSLATE_NOOP("Missions", "The Timber Levy"),
     QT_TRANSLATE_NOOP("Missions", "The allied wing is at your works. Italians, mostly. They fight for Rome because the alternative is fighting Rome, which is a very persuasive argument."),
     QT_TRANSLATE_NOOP("Missions", "The allied wing is down to a handful. I have counted them twice, because there was time. Whatever happens next on this field, remember someone counted."),
@@ -238,11 +257,14 @@ namespace {
     QT_TRANSLATE_NOOP("Missions", "The commander falls and the line behind him is just men standing in a dry riverbed."),
     QT_TRANSLATE_NOOP("Missions", "The commander falls — and every line collapses behind him."),
     QT_TRANSLATE_NOOP("Missions", "The commander falls — the vigil ends. And the dead keep walking."),
+    QT_TRANSLATE_NOOP("Missions", "The consul falls and the gates answer to nobody."),
     QT_TRANSLATE_NOOP("Missions", "The crew is gone and the officer is standing in a clearing with an axe. That is not a work party."),
     QT_TRANSLATE_NOOP("Missions", "The crossing is a running battle against terrain, local resistance, and Roman influence. Sepulcher grave-lights remain a dangerous side path rather than the campaign's main enemy."),
     QT_TRANSLATE_NOOP("Missions", "The dead of the Iron Sepulcher still contest the field; keep a reserve."),
+    QT_TRANSLATE_NOOP("Missions", "The fields are across the river and the roads are still open. Send the crews out now; they will not be open long."),
     QT_TRANSLATE_NOOP("Missions", "The final confrontation is Carthage against three coordinated Roman commands, until an Iron Sepulcher host attacks from the battlefield's neglected flank."),
     QT_TRANSLATE_NOOP("Missions", "The final pass is nearly cleared. Whatever comes down the far side, I shall be in Italy waiting for it with a great many more men than this."),
+    QT_TRANSLATE_NOOP("Missions", "The granges on the south bank are the nearest stockpiles to the grain."),
     QT_TRANSLATE_NOOP("Missions", "The high pass is sending everything it has. They have been paid in Roman silver to hold that road, and mountain men earn their silver."),
     QT_TRANSLATE_NOOP("Missions", "The hill fort is yours. It was built to watch a river; now it will watch you cross one. Forts do not much care who is inside them."),
     QT_TRANSLATE_NOOP("Missions", "The last camp lies to the north-east, beyond the consular line."),
@@ -259,6 +281,7 @@ namespace {
     QT_TRANSLATE_NOOP("Missions", "The northern legion is coming down through the orchards. Slowly. I ordered slowly. Every tree they pass is one you will not harvest."),
     QT_TRANSLATE_NOOP("Missions", "The northern legion is spent. I have others. I have always had others. That is what the delaying was for."),
     QT_TRANSLATE_NOOP("Missions", "The officer of the detail falls. Nobody else here can sign for the levy."),
+    QT_TRANSLATE_NOOP("Missions", "The old Republican circuit and the citadel are still walls. Falling back is not losing."),
     QT_TRANSLATE_NOOP("Missions", "The pass keeps them, then. A word of advice to whoever reads this: do not follow the lights on the snowfield. Whatever carries a lantern at that height stopped being a shepherd a long time ago."),
     QT_TRANSLATE_NOOP("Missions", "The patrols are nearly finished. So be it. I shall take the fleet back to Italy and wait for you there with the rest of the Republic."),
     QT_TRANSLATE_NOOP("Missions", "The quarter holds when the last column is dead in the streets."),
@@ -289,8 +312,11 @@ namespace {
     QT_TRANSLATE_NOOP("Missions", "There. That is what the river was for. Have your surgeons work quickly, and tell whatever is left of your baggage train that the road home goes back up the mountain."),
     QT_TRANSLATE_NOOP("Missions", "They are getting up again. Do not look at the faces; you will know some of them. Fall back to the road, keep walking, and do not stop to bury anything. It does not stay buried here."),
     QT_TRANSLATE_NOOP("Missions", "They are hitting your camp, because it is the only thing on this road they can see. Men in fog attack whatever is nearest. Remember it."),
+    QT_TRANSLATE_NOOP("Missions", "They are on a gate. A gate is one building and it is the only thing between them and a street they can walk down. Answer it or fall back to the old circuit; do not do half of each."),
     QT_TRANSLATE_NOOP("Missions", "They tell me you have taken an army into the Alps in October, with elephants. I have stopped chasing you. The mountain is a more thorough officer than I am and it does not require pay."),
     QT_TRANSLATE_NOOP("Missions", "Three Roman columns have entered the consular roads. The first coordinated assault arrives at 17:00."),
+    QT_TRANSLATE_NOOP("Missions", "Three walls, seventeen gates, and not one field inside them. The capital cannot be held by standing on its ramparts and it cannot be fed without leaving them. Break six assaults and still bring the harvest in."),
+    QT_TRANSLATE_NOOP("Missions", "Three walls. We only ever needed the fields. Open the granaries and put out the fires on the sacred way -- I want the citadel standing when they come to see what is left of it."),
     QT_TRANSLATE_NOOP("Missions", "Timber for the sledges, stone for the roadbed, iron for shoes and axles."),
     QT_TRANSLATE_NOOP("Missions", "Timber only counts once a builder unloads it at the camp. The richest stands are the furthest from the stockpile, so the mission is won on the length of the walk, not the swing of the axe."),
     QT_TRANSLATE_NOOP("Missions", "Turn the nearest Roman allied camp south of the field."),
@@ -300,8 +326,11 @@ namespace {
     QT_TRANSLATE_NOOP("Missions", "Varro's town. Varro is not in it; Varro is out front, being brave at other people. Enjoy the walls. I never needed them."),
     QT_TRANSLATE_NOOP("Missions", "We are losing men on the goat path. Pull the wounded back and put fresh spears in the gap. Nobody dies standing in a hole."),
     QT_TRANSLATE_NOOP("Missions", "We are on your bank and at your works. A river only stops men who are not already freezing."),
+    QT_TRANSLATE_NOOP("Missions", "We are spending the garrison faster than the city makes it. Pull back inside the Republican wall. It is shorter, it is older, and it has four gates instead of nine."),
+    QT_TRANSLATE_NOOP("Missions", "We will not climb your wall, consul. We will stand on your fields and wait, and you will open the gate yourself to go and cut your own wheat."),
     QT_TRANSLATE_NOOP("Missions", "With the river camp taken there is nowhere left for the legions to reform."),
     QT_TRANSLATE_NOOP("Missions", "You bent us and we did not break, and that is the only sentence anyone will need. Gather your dead off the Aufidus. The river runs slow here; they will be a long time going by."),
+    QT_TRANSLATE_NOOP("Missions", "You cannot cover every gate. Watch the telegraph and move on interior lines."),
     QT_TRANSLATE_NOOP("Missions", "You have a walled town, an orchard, and a very great many mouths. I have the roads. I am not going to fight you. I am going to stand between you and every granary in Campania until winter explains my position better than I can."),
     QT_TRANSLATE_NOOP("Missions", "You have the allied camp. Socii - allies. They were promised land for this war, and you have just handed them the first piece of it face down."),
     QT_TRANSLATE_NOOP("Missions", "You have the marching town. Flaminius built it in a fog and never looked back. There is a lesson there; he will not be the one to learn it."),
@@ -340,12 +369,14 @@ namespace {
     QT_TRANSLATE_NOOP("Campaigns", "The last standards fall under an African sky. Rome and Carthage will remember the field differently; the Iron Sepulcher remembers only that both armies finally entered its reach."),
     QT_TRANSLATE_NOOP("Campaigns", "The survivors flee toward Placentia through sleet and mud. Roman tents, grain, and weapons are captured; bodies beneath the ice rise only when nobody is watching."),
 
-    // ---- Maps (28) ----
+    // ---- Maps (30) ----
     QT_TRANSLATE_NOOP("Maps", "A fertile valley for two. One channel runs the whole map north to south and only two fords cross it. A cross spur on each bank means you either hug the water or swing wide to reach them, and the boomerang ridge cupping each start makes the back door terrain rather than a wall. Nothing haunts this valley: it is the plain match, with no Sepulcher to distract either side."),
     QT_TRANSLATE_NOOP("Maps", "A quiet meadow between a Carthaginian camp and a Roman outpost, laid out for the tutorial: timber, stone and ore within a short walk of the barracks, a scouting party to practise on, and a small enemy camp to take once the army is ready."),
     QT_TRANSLATE_NOOP("Maps", "A redoubt above a dry ford, for one. A spur of rock runs the length of the approach and the two shoulders close the flanks, so everything coming down from the north is committed to one of two lanes -- the cart gate in the west, the goat path in the east -- and they do not meet again until the ground in front of the camp. One line cannot hold both mouths, so holding here is a question of which lane you stand in and when you leave it."),
     QT_TRANSLATE_NOOP("Maps", "A timber detail's cut above the Pinewater, for one. The near stand beside the camp is already half felled, so the levy is only filled out of the deep pines a long haul north-east, past the spine of rock the haul road has to swing around. The scree that gives up the stone lies the other way entirely, and the crew is not big enough to work both without splitting it."),
     QT_TRANSLATE_NOOP("Maps", "Amber Delta"),
+    QT_TRANSLATE_NOOP("Maps", "An enormous capital on a 768 m stage, built in rings. Farmland, orchards, quarries, lakes and a ruined old town fill the countryside; a great eight-gated wall encloses the new city with its circus, theatre, docks and insulae; an older Republican circuit holds the forum and the ancient core; and a walled citadel crowns the sacred mountain behind the whole thing, where the healers keep the temples."),
+    QT_TRANSLATE_NOOP("Maps", "Aurelia Magna"),
     QT_TRANSLATE_NOOP("Maps", "Battle of Cannae"),
     QT_TRANSLATE_NOOP("Maps", "Battle of Lake Trasimene"),
     QT_TRANSLATE_NOOP("Maps", "Battle of Ticino"),
